+++ superwall/src/main/java/com/superwall/sdk/logger/LogScope.kt
@@ -4,6 +4,7 @@ enum class LogScope {
     localizationManager,
     bounceButton,
     webEntitlements,
+    customerInfo,
     coreData,
     configManager,
     identityManager,