+++ CHANGELOG.md
@@ -5,8 +5,10 @@ The changelog for `Superwall`. Also see the [releases](https://github.com/superw
 ## 2.6.6
 
 ## Enhancements
-
 - Add dynamic notification support and scheduling enabling deeper personalization of notifications 
+- Provides a `CustomerInfo` class API, allowing you to observe the customer's purchases and subscription lifecycle via `Superwall.instance.customerInfo` flow
+- Provides a new delegate method to observe customer info changes - `fun customerInfoDidChange(from: CustomerInfo, to: CustomerInfo)`  
+- Provides a `CustomerInfoDidChange` event to track customer info changes
 
 ## 2.6.5